stetson_server:message - update spec
Our project is giving us dialyzer warnings when we call stetson:timer
with a string().  stetson:timer's spec says this is okay, and indeed it works,
but when it tries to pass along this value to stetson_server:cast, it triggers
a dialyzer failure because stetson_server:message expects only an atom there.